Handle node split edge case in react plugin.
In some scenarios, it's possible to insert a node such that the node
with the lower position is the new one, and the one with the higher
position is the old/pre-existing one. In this case, we now maintain key
persistence properly, which resolves an issue where the portal registry
can end up with multiple portals with the same key.